+++ app/Transformers/AccountTransformer.php
@@ -114,7 +114,7 @@ class AccountTransformer extends AbstractTransformer
             'opening_balance'         => $openingBalance,
             'opening_balance_date'    => $openingBalanceDate,
             'liability_type'          => $liabilityType,
-            'interest'                => (float)$interest,
+            'interest'                => $interest,
             'interest_period'         => $interestPeriod,
             'include_net_worth'       => $includeNetWorth,
             'longitude'               => $longitude,