+++ frontend/src/components/dashboard/TopBoxes.vue
@@ -30,7 +30,7 @@
           <span v-if="error" class="info-box-text"><i class="fas fa-exclamation-triangle text-danger"></i></span>
           <!-- balance in preferred currency -->
           <span v-for="balance in prefCurrencyBalances" :title="balance.sub_title" class="info-box-number">{{ balance.value_parsed }}</span>
-
+          <span v-if="0 === prefCurrencyBalances.length" class="info-box-number">&nbsp;</span>
           <div class="progress bg-info">
             <div class="progress-bar" style="width: 0"></div>
           </div>
@@ -55,6 +55,7 @@
           <span v-if="error" class="info-box-text"><i class="fas fa-exclamation-triangle text-danger"></i></span>
           <!-- bills unpaid, in preferred currency. -->
           <span v-for="balance in prefBillsUnpaid" class="info-box-number">{{ balance.value_parsed }}</span>
+          <span v-if="0===prefBillsUnpaid.length" class="info-box-number">&nbsp;</span>
 
           <div class="progress bg-teal">
             <div class="progress-bar" style="width: 0"></div>
@@ -80,6 +81,7 @@
           <span v-if="error" class="info-box-text"><i class="fas fa-exclamation-triangle text-danger"></i></span>
           <!-- left to spend in preferred currency -->
           <span v-for="left in prefLeftToSpend" :title="left.sub_title" class="info-box-number">{{ left.value_parsed }}</span>
+          <span v-if="0 === prefLeftToSpend.length" class="info-box-number">&nbsp;</span>
 
           <div class="progress bg-success">
             <div class="progress-bar" style="width: 0"></div>
@@ -105,7 +107,7 @@
           <span v-if="loading && !error" class="info-box-text"><i class="fas fa-spinner fa-spin"></i></span>
           <span v-if="error" class="info-box-text"><i class="fas fa-exclamation-triangle text-danger"></i></span>
           <span v-for="nw in prefNetWorth" :title="nw.sub_title" class="info-box-number">{{ nw.value_parsed }}</span>
-
+          <span v-if="0===prefNetWorth.length">&nbsp;</span>
           <div class="progress bg-success">
             <div class="progress-bar" style="width: 0"></div>
           </div>
@@ -124,6 +126,7 @@
 
 <script>
 import {createNamespacedHelpers} from "vuex";
+import format from 'date-fns/format';
 
 const {mapState, mapGetters, mapActions, mapMutations} = createNamespacedHelpers('dashboard/index')
 export default {
@@ -249,8 +252,12 @@ export default {
       this.billsUnpaid = [];
       this.leftToSpend = [];
       this.netWorth = [];
-      let startStr = this.start.toISOString().split('T')[0];
-      let endStr = this.end.toISOString().split('T')[0];
+      let startStr = format(this.start, 'y-MM-dd');
+      let endStr = format(this.end, 'y-MM-dd');
+      //let startStr = this.start.toISOString().split('T')[0];
+      //let endStr = this.end.toISOString().split('T')[0];
+      //console.log(startStr);
+      //console.log(endStr);
       axios.get('./api/v1/summary/basic?start=' + startStr + '&end=' + endStr)
           .then(response => {
             this.summary = response.data;