+++ resources/lang/sk_SK/validation.php
@@ -134,8 +134,8 @@ return [
     'starts_with'                    => 'Hodnota musí začínať :values.',
     'unique_webhook'                 => 'Webhook s týmito hodnotami už existuje.',
     'unique_existing_webhook'        => 'Iný webhook s týmito hodnotami už existuje.',
-    'same_account_type'              => 'Both accounts must be of the same account type',
-    'same_account_currency'          => 'Both accounts must have the same currency setting',
+    'same_account_type'              => 'Oba účty musia mať rovnaký typ',
+    'same_account_currency'          => 'Oba účty musia mať rovnakú menu',
 
     'secure_password'             => 'Toto nie je bezpečné heslo. Skúste iné. Viac se dozviete na http://bit.ly/FF3-password-security',
     'valid_recurrence_rep_type'   => 'Neplatný typ opakovania pre opakované transakcie.',
@@ -185,7 +185,7 @@ return [
     'withdrawal_dest_need_data'   => 'Pro pokračovanie je potrebné platné ID zdrojového účtu a/alebo platný názov zdrojového účtu.',
     'withdrawal_dest_bad_data'    => 'Pre ID „:id“ alebo mena „:name“ sa nenašiel žiadny platný cieľový účet.',
 
-    'generic_source_bad_data'  => 'Could not find a valid source account when searching for ID ":id" or name ":name".',
+    'generic_source_bad_data'  => 'Pre ID ":id" alebo meno ":name" sa nenašiel žiadny platný zdrojový účet.',
 
     'deposit_source_need_data' => 'Pre pokračovanie je potrebné platné ID zdrojového účtu a/alebo platný názov zdrojového účtu.',
     'deposit_source_bad_data'  => 'Pre ID „:id“ alebo „:name“ sa nenašiel žiadny platný zdrojový účet.',